+++ Cargo.toml
@@ -98,6 +98,8 @@ quantette = { version = "0.3.0", features = ["threads"] }
 sysinfo = "0.33.1"
 dicom-pixeldata = { version = "0.8.0", features = ["image"] }
 dicom-object = "0.8.0"
+unicode-segmentation = "1.12.0"
+font-kit = "0.14.2"
 
 [features]
 default = [
